+++ node/resolvers/postResolvers.ts
@@ -1,4 +1,7 @@
+/* eslint-disable no-await-in-loop */
 /* eslint-disable @typescript-eslint/camelcase */
+const API_MAX_RETURN = 100
+
 export const postResolvers = {
   author: async (
     { author }: { author: number },
@@ -23,7 +26,31 @@ export const postResolvers = {
     const {
       clients: { wordpressProxy },
     } = ctx
-    return categories.map(id => wordpressProxy.getCategory(id, customDomain))
+
+    let page = 1
+    let returned = 0
+    const allCategories: WpCategory[] = []
+
+    try {
+      do {
+        const { data } = await wordpressProxy.getCategories({
+          customDomain,
+          page,
+          per_page: API_MAX_RETURN,
+          include: categories,
+        })
+
+        if (data) {
+          allCategories.push(...data)
+        }
+
+        returned = data?.length || 0
+        page += 1
+      } while (returned === API_MAX_RETURN)
+    } catch (err) {
+      return []
+    }
+    return allCategories
   },
   tags: async (
     { tags }: { tags: [number] },
@@ -33,7 +60,31 @@ export const postResolvers = {
     const {
       clients: { wordpressProxy },
     } = ctx
-    return tags.map(id => wordpressProxy.getTag(id, customDomain))
+
+    let page = 1
+    let returned = 0
+    const allTags: WpTag[] = []
+
+    try {
+      do {
+        const { data } = await wordpressProxy.getTags({
+          customDomain,
+          page,
+          per_page: API_MAX_RETURN,
+          include: tags,
+        })
+
+        if (data) {
+          allTags.push(...data)
+        }
+
+        returned = data?.length || 0
+        page += 1
+      } while (returned === API_MAX_RETURN)
+    } catch (err) {
+      return []
+    }
+    return allTags
   },
   featured_media: async (
     { featured_media }: { featured_media: number },